Fix: Refactored calculator functions into individual files.
Fix: Refactored calculator functions into individual files.
