+++ Calculator/multiplication.py
@@ -1,0 +1,3 @@
+def multiplication(a, b):
+    c = float(a)*float(b)
+    return float(c)
