+++ Calculator/squared.py
@@ -1,0 +1,3 @@
+def squared(a):
+    c = float(a) ** 2
+    return c
